--- components/common/ExpandableSearchGrid/README.md
@@ -1,5 +1,0 @@
-### Expandable Search Grid Example
-
-```ts
-<ExpandableSearchGrid />
-```
