--- components/common/SplashPageSearchBar/README.md
@@ -1,5 +1,0 @@
-### Splash Page Search Example
-
-```ts
-<SplashPageSearchBar />
-```
