+++ pages/dashboard/index.tsx
@@ -14,8 +14,8 @@ import type { NextPage } from 'next';
 import Head from 'next/head';
 import React, { useCallback, useEffect, useState } from 'react';
 
+import SearchBar from '../../components/common/SearchBar/searchBar';
 import Carousel from '../../components/common/Carousel/carousel';
-import { ExpandableSearchGrid } from '../../components/common/ExpandableSearchGrid/expandableSearchGrid';
 import ProfessorCard from '../../components/common/ProfessorCard/ProfessorCard';
 import { RelatedClasses } from '../../components/common/RelatedClasses/relatedClasses';
 import { BarGraph } from '../../components/graph/BarGraph/BarGraph';
@@ -671,12 +671,10 @@ export const Dashboard: NextPage = () => {
       </Head>
       <div className=" w-full bg-light h-full">
         <TopMenu />
-        <ExpandableSearchGrid
-          onChange={searchTermsChange}
-          setIncluded={setIncluded}
-          studentTotals={studentTotals}
-          relatedQuery={relatedQuery}
-          averageData={averageData}
+        <SearchBar
+          manageQuery={true}
+          path={'/dashboard'}
+          changeValue={(val: SearchQuery[]) => console.log(val)}
         />
         <div className="w-full h-5/6 justify-center">
           <div className="w-full h-5/6 relative min-h-full">
